refactor: migrate extensions commands to core24
