+++ snapcraft/commands/extensions.py
@@ -1,6 +1,6 @@
 # -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
 #
-# Copyright 2024 Canonical Ltd.
+# Copyright 2022-2024 Canonical Ltd.
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License version 3 as
@@ -14,16 +14,126 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Extension commands for core24 (forwarded to the shared core22 implementation)."""
+"""Snapcraft extension commands."""
 
+import textwrap
+from typing import Dict, List
+
+import tabulate
+import yaml
 from craft_application.commands import AppCommand
+from craft_cli import emit
+from overrides import overrides
+from pydantic import BaseModel
 
-from snapcraft.commands import core22
+from snapcraft import extensions, models
+from snapcraft.parts.yaml_utils import (
+    apply_yaml,
+    extract_parse_info,
+    get_snap_project,
+    process_yaml,
+)
+from snapcraft.utils import get_host_architecture
+from snapcraft_legacy.internal.project_loader import (
+    find_extension,
+    supported_extension_names,
+)
 
 
-class ExpandExtensions(AppCommand, core22.ExpandExtensionsCommand):
-    """core24 command to expand extensions."""
+class ExtensionModel(BaseModel):
+    """Extension model for presentation."""
+
+    name: str
+    bases: List[str]
+
+    def marshal(self) -> Dict[str, str]:
+        """Marshal model into a dictionary for presentation."""
+        return {
+            "Extension name": self.name,
+            "Supported bases": ", ".join(sorted(self.bases)),
+        }
 
 
-class ListExtensions(AppCommand, core22.ListExtensionsCommand):
-    """core24 command to list extensions."""
+class ListExtensionsCommand(AppCommand):
+    """List available extensions for all supported bases."""
+
+    name = "list-extensions"
+    help_msg = "List available extensions for all supported bases."
+    overview = textwrap.dedent(
+        """
+        List available extensions and their corresponding bases.
+        """
+    )
+
+    @overrides
+    def run(self, parsed_args):
+        extension_presentation: Dict[str, ExtensionModel] = {}
+
+        # New extensions.
+        for extension_name in extensions.registry.get_extension_names():
+            extension_class = extensions.registry.get_extension_class(extension_name)
+            extension_bases = list(extension_class.get_supported_bases())
+            extension_presentation[extension_name] = ExtensionModel(
+                name=extension_name, bases=extension_bases
+            )
+
+        # Extensions from snapcraft_legacy.
+        for _extension_name in supported_extension_names():
+            extension_class = find_extension(_extension_name)
+            extension_name = _extension_name.replace("_", "-")
+            extension_bases = list(extension_class.get_supported_bases())
+            if extension_name in extension_presentation:
+                extension_presentation[extension_name].bases += extension_bases
+            else:
+                extension_presentation[extension_name] = ExtensionModel(
+                    name=extension_name, bases=extension_bases
+                )
+
+        printable_extensions = sorted(
+            [v.marshal() for v in extension_presentation.values()],
+            key=lambda d: d["Extension name"],
+        )
+        emit.message(tabulate.tabulate(printable_extensions, headers="keys"))
+
+
+class ExtensionsCommand(ListExtensionsCommand):
+    """A command alias to list the available extensions."""
+
+    name = "extensions"
+    hidden = True
+
+
+class ExpandExtensionsCommand(AppCommand):
+    """Expand the extensions in the snapcraft.yaml file."""
+
+    name = "expand-extensions"
+    help_msg = "Expand extensions in snapcraft.yaml"
+    overview = textwrap.dedent(
+        """
+        Extensions selected in apps in snapcraft.yaml will be
+        expanded and shown as output.
+        """
+    )
+
+    @overrides
+    def run(self, parsed_args):
+        snap_project = get_snap_project()
+
+        # load yaml file and trigger legacy behavior if base is core, core18, or core20
+        yaml_data = process_yaml(snap_project.project_file)
+
+        # process yaml before unmarshalling the data
+        arch = get_host_architecture()
+        yaml_data_for_arch = apply_yaml(yaml_data, arch, arch)
+
+        # `apply_yaml()` adds or replaces the architectures keyword with an Architecture
+        # object, which does not easily dump to a yaml file
+        yaml_data_for_arch.pop("architectures", None)
+        yaml_data_for_arch.pop("platforms", None)
+
+        # `parse-info` keywords must be removed before unmarshalling, because they are
+        # not part of the Project model
+        extract_parse_info(yaml_data_for_arch)
+
+        models.Project.unmarshal(yaml_data_for_arch)
+        emit.message(yaml.safe_dump(yaml_data_for_arch, indent=4, sort_keys=False))
