+++ tests/unit/commands/test_list_extensions.py
@@ -1,6 +1,6 @@
 # -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
 #
-# Copyright 2022 Canonical Ltd.
+# Copyright 2022,2024 Canonical Ltd.
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License version 3 as
@@ -19,16 +19,15 @@ from textwrap import dedent
 
 import pytest
 
-import snapcraft.commands.core22
+import snapcraft.commands
 
 
 @pytest.mark.usefixtures("fake_extension")
 @pytest.mark.parametrize(
     "command",
     [
-        snapcraft.commands.core22.ListExtensionsCommand,
-        snapcraft.commands.core22.ExtensionsCommand,
-        snapcraft.commands.ListExtensions,
+        snapcraft.commands.ListExtensionsCommand,
+        snapcraft.commands.ExtensionsCommand,
     ],
 )
 def test_command(emitter, command):
@@ -73,9 +72,8 @@ def test_command(emitter, command):
 @pytest.mark.parametrize(
     "command",
     [
-        snapcraft.commands.core22.ListExtensionsCommand,
-        snapcraft.commands.core22.ExtensionsCommand,
-        snapcraft.commands.ListExtensions,
+        snapcraft.commands.ListExtensionsCommand,
+        snapcraft.commands.ExtensionsCommand,
     ],
 )
 def test_command_extension_dups(emitter, command):
